add fluent validation

+++ GymManagementApi/Model/PaymentModel.cs
@@ -6,6 +6,6 @@
         public string PaymentMethod { get; set; }
         public string MemberName { get; set; }
         public DateTime PaymentDate { get; set; }
-        public double MemberID { get; set; }
+        public int MemberID { get; set; }
     }
 }
